+++ kernel/shell/shell_internal.h
@@ -510,6 +510,7 @@ void CmdLsgpt();
 void CmdLsmod();
 void CmdMkfs(u32 argc, char** argv);
 void CmdMkfsDuetfs(u32 argc, char** argv);
+void CmdInstall(u32 argc, char** argv);
 void CmdLastdump();
 
 // ---------------------------------------------------------------